// src/game/data/enums.ts
export const enum EAddonType {
  NONE = 0,
  SLOT = 1,
  STAT = 2,
  ABILITY_POINT = 3,
  COMBAT_EFFECT = 4,
  SKILL_DAMAGE = 5,
  SKILL_CRITICAL_RATIO = 6,
  SKILL_CRITICAL_DAMAGE = 7,
  SKILL_PENETRATION = 8,
  NPC_GRADE_LESS_DAMAGE = 9,
  NPC_GRADE_LESS_CRITICAL_RATIO = 10,
  NPC_GRADE_LESS_CRITICAL_DAMAGE = 11,
  NPC_GRADE_LESS_PENETRATION = 12,
  NPC_GRADE_GREATER_DAMAGE = 13,
  NPC_GRADE_GREATER_CRITICAL_RATIO = 14,
  NPC_GRADE_GREATER_CRITICAL_DAMAGE = 15,
  NPC_GRADE_GREATER_PENETRATION = 16,
  NPC_SPECIES_DAMAGE = 17,
  NPC_SPECIES_CRITICAL_RATIO = 18,
  NPC_SPECIES_CRITICAL_DAMAGE = 19,
  NPC_SPECIES_PENETRATION = 20,
  NPC_ATTR_DAMAGE = 21,
  NPC_ATTR_CRITICAL_RATIO = 22,
  NPC_ATTR_CRITICAL_DAMAGE = 23,
  NPC_ATTR_PENETRATION = 24,
  MANA_REDUCTION = 25,
  SKILL_MANA_REDUCTION = 26,
  SKILL_COOLDOWN_REDUCTION = 27,
  ABILITY_FEATURE = 28,
  CLASS_OPTION = 29,
  ABILITY_POINT_PASSIVE = 30,
  INSTRUMENT = 31,
  SKILL_FEATURE = 32,
  NPC_ADAPTATION = 33,
  SKILL_GROUP_DAMAGE = 34,
  SKILL_GROUP_COOLDOWN_REDUCTION = 35,
  SKILL_LEVEL = 36,
  SKILL_FEATURE_LEVEL = 37,
  LIFE_CASTING_SPEED = 38,
  LIFE_CASTING_TIER = 39,
  LIFE_BONUS_TYPE_SUCCESS = 40,
  LIFE_BONUS_TYPE_EXTRA = 41,
  LIFE_BONUS_TYPE_SKILL_BONUS = 42,
  LIFE_BONUS_TYPE_MINIGAME_PERFECT = 43,
  LIFE_DURABILITY_BONUS = 44,
  LIFE_MINI_GAME_DIFFICULTY = 45,
  COMBAT_EFFECT_COOLDOWN_REDUCTION = 46,
  SKILL_DAMAGE_ADDEND = 47,
  AWAKENING_USABLE_COUNT_ADDEND = 48,
  BATTLE_ITEM_HEAL = 49,
  PARTY_HEAL = 50,
  PARTY_SHIELD = 51,
  IDENTITY_GAUGE = 52,
  ATTACK_POWER_AMPLIFY_ADDEND = 53,
  ATTACK_POWER_AMPLIFY_MULTIPLIER = 54,
  NOT_IN_PARTY_DAMAGE = 55,
  SKILL_EFFECT_GROUP_SET_DAMAGE = 56,
  ARK_PASSIVE_POINT = 57,
  SKILL_STATUS_EFFECT_STAT_MULTIPLIER = 58,
  SKILL_GROUP_STATUS_EFFECT_STAT_MULTIPLIER = 59,
  SKILL_PARTY_SHIELD = 60,
  SKILL_GROUP_PARTY_SHIELD = 61,
  SKILL_PARTY_HEAL = 62,
  SKILL_GROUP_PARTY_HEAL = 63,
  SKILL_ATTACK_POWER_AMPLIFY_MULTIPLIER = 64,
  SKILL_GROUP_ATTACK_POWER_AMPLIFY_MULTIPLIER = 65,
  LIFE_BONUS_TYPE_COMMON = 66,
  LIFE_BONUS_TYPE_UNCOMMON = 67,
  LIFE_BONUS_TYPE_RARE = 68,
  LIFE_BONUS_TYPE_EPIC = 69,
  LIFE_BONUS_TYPE_LEGEND = 70,
  LIFE_BONUS_TYPE_ALL = 71,
  LIFE_BONUS_TYPE_MINIGAME = 72,
  LIFE_BONUS_TYPE_CHEST = 73,
  LIFE_SPECIAL_SKILL_CHARGE_BONUS = 74,
  LIFE_HERBALISM_SPECIAL_SKILL_BUFF_TIME = 75,
  LIFE_TREASURE_HUNTER_BONUS = 76
}

export const enum EStatType {
  NONE = 0,
  HP = 1,
  MP = 2,
  STR = 3,
  AGI = 4,
  INT = 5,
  CON = 6,
  STR_X = 7,
  AGI_X = 8,
  INT_X = 9,
  CON_X = 10,
  CRITICALHIT = 15,
  SPECIALTY = 16,
  OPPRESSION = 17,
  RAPIDITY = 18,
  ENDURANCE = 19,
  MASTERY = 20,
  CRITICALHIT_X = 21,
  SPECIALTY_X = 22,
  OPPRESSION_X = 23,
  RAPIDITY_X = 24,
  ENDURANCE_X = 25,
  MASTERY_X = 26,
  MAX_HP = 27,
  MAX_MP = 28,
  MAX_HP_X = 29,
  MAX_MP_X = 30,
  MAX_HP_X_X = 31,
  MAX_MP_X_X = 32,
  NORMAL_HP_RECOVERY = 33,
  COMBAT_HP_RECOVERY = 34,
  NORMAL_HP_RECOVERY_RATE = 35,
  COMBAT_HP_RECOVERY_RATE = 36,
  NORMAL_MP_RECOVERY = 37,
  COMBAT_MP_RECOVERY = 38,
  NORMAL_MP_RECOVERY_RATE = 39,
  COMBAT_MP_RECOVERY_RATE = 40,
  SELF_RECOVERY_RATE = 41,
  DRAIN_HP_DAM_RATE = 42,
  DRAIN_MP_DAM_RATE = 43,
  DAM_REFLECTION_RATE = 44,
  EVOLUTION_DAM_RATE = 45,
  STIGMA_POWER_RATE = 46,
  CHAR_ATTACK_DAM = 47,
  SKILL_EFFECT_DAM_ADDEND = 48,
  ATTACK_POWER_RATE = 49,
  SKILL_DAMAGE_RATE = 50,
  ATTACK_POWER_RATE_X = 51,
  SKILL_DAMAGE_RATE_X = 52,
  COOLDOWN_REDUCTION = 53,
  PARALYZATION_POINT_RATE = 54,
  DEF = 55,
  RES = 56,
  DEF_X = 57,
  RES_X = 58,
  DEF_X_X = 59,
  RES_X_X = 60,
  CRITICAL_PHYSICAL_INC_RATE = 65,
  CRITICAL_MAGICAL_INC_RATE = 66,
  DEF_PEN_RATE = 67,
  RES_PEN_RATE = 68,
  PHYSICAL_INC_RATE = 69,
  MAGICAL_INC_RATE = 70,
  SELF_SHIELD_RATE = 71,
  HIT_RATE = 72,
  DODGE_RATE = 73,
  CRITICAL_HIT_RATE = 74,
  CRITICAL_RES_RATE = 75,
  CRITICAL_DAM_RATE = 76,
  ATTACK_SPEED = 77,
  ATTACK_SPEED_RATE = 78,
  MOVE_SPEED = 79,
  MOVE_SPEED_RATE = 80,
  PROP_MOVE_SPEED = 81,
  PROP_MOVE_SPEED_RATE = 82,
  VEHICLE_MOVE_SPEED = 83,
  VEHICLE_MOVE_SPEED_RATE = 84,
  SHIP_MOVE_SPEED = 85,
  SHIP_MOVE_SPEED_RATE = 86,
  FIRE_DAM_RATE = 87,
  ICE_DAM_RATE = 88,
  ELECTRICITY_DAM_RATE = 89,
  EARTH_DAM_RATE = 91,
  DARK_DAM_RATE = 92,
  HOLY_DAM_RATE = 93,
  ELEMENTS_DAM_RATE = 94,
  FIRE_RES_RATE = 95,
  ICE_RES_RATE = 96,
  ELECTRICITY_RES_RATE = 97,
  EARTH_RES_RATE = 99,
  DARK_RES_RATE = 100,
  HOLY_RES_RATE = 101,
  ELEMENTS_RES_RATE = 102,
  ULTIMATE_AWAKENING_DAM_RATE = 104,
  SELF_CC_TIME_RATE = 105,
  ENEMY_CC_TIME_RATE = 106,
  IDENTITY_VALUE1 = 107,
  IDENTITY_VALUE2 = 108,
  IDENTITY_VALUE3 = 109,
  AWAKENING_DAM_RATE = 110,
  ITEM_DROP_RATE = 111,
  GOLD_RATE = 112,
  EXP_RATE = 113,
  ATTACK_POWER_ADDEND = 123,
  ATTACK_POWER_ADDEND_2 = 124,
  NPC_SPECIES_HUMANOID_DAM_RATE = 125,
  NPC_SPECIES_DEVIL_DAM_RATE = 126,
  NPC_SPECIES_SUBSTANCE_DAM_RATE = 127,
  NPC_SPECIES_UNDEAD_DAM_RATE = 128,
  NPC_SPECIES_PLANT_DAM_RATE = 129,
  NPC_SPECIES_INSECT_DAM_RATE = 130,
  NPC_SPECIES_SPIRIT_DAM_RATE = 131,
  NPC_SPECIES_WILD_BEAST_DAM_RATE = 132,
  NPC_SPECIES_MECHANIC_DAM_RATE = 133,
  NPC_SPECIES_ANCIENT_DAM_RATE = 134,
  NPC_SPECIES_GOD_DAM_RATE = 135,
  NPC_SPECIES_ARCHFIEND_DAM_RATE = 136,
  VITALITY = 137,
  SHIP_BOOTER_SPEED = 138,
  SHIP_WRECK_SPEED_RATE = 139,
  ISLAND_SPEED_RATE = 140,
  ATTACK_POWER_SUB_RATE_1 = 141,
  ATTACK_POWER_SUB_RATE_2 = 142,
  PHYSICAL_INC_SUB_RATE_1 = 143,
  PHYSICAL_INC_SUB_RATE_2 = 144,
  MAGICAL_INC_SUB_RATE_1 = 145,
  MAGICAL_INC_SUB_RATE_2 = 146,
  SKILL_DAMAGE_SUB_RATE_1 = 147,
  SKILL_DAMAGE_SUB_RATE_2 = 148,
  RESOURCE_RECOVERY_RATE = 149,
  BASE_DAMAGE_RATE = 150,
  WEAPON_DAM = 151,
  WEAPON_DAM_X = 152
}